--- app/controllers/candidate/base_controller.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Candidate
-  class BaseController < ApplicationController
-    before_action :authenticate_employee!
-
-    alias current_user current_employee
-  end
-end
