+++ app/policies/response_policy.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class ResponsePolicy < ApplicationPolicy
+  def new?
+    user.is_a? Employee
+  end
+
+  def create?
+    user.is_a? Employee
+  end
+end
