+++ config/locales/ru.yml
@@ -42,4 +42,3 @@ ru:
       phone: "Контактный телефон"
       resume: "Резюме"
       covering_letter: "Сопроводительное письмо"
-  not_allowed: "не разрешен доступ к этому действию"
