+++ config/routes.rb
@@ -14,10 +14,8 @@ Rails.application.routes.draw do
     end
   end
 
-  namespace :candidate do
-    resources :vacancies, shallow: true do
-      resources :responses, only: %i[show new create]
-    end
+  resources :vacancies, shallow: true, only: %i[index show] do
+    resources :responses, only: %i[show new create]
   end
