+++ postcss.config.js
@@ -1,4 +1,4 @@
-lmodule.exports = {
+module.exports = {
   plugins: [
     require('postcss-import'),
     require('postcss-flexbugs-fixes'),
